back/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
boto3==1.42.81
botocore==1.42.81
certifi==2026.2.25
cffi==2.0.0
cryptography==46.0.7
fastapi==0.135.2
google-auth==2.49.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
passlib==1.7.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
